Move flycam "player" logic into its own module
