+++ src/flycam.rs
@@ -1,0 +1,94 @@
+//! This module contains everything about the controls and rendering related
+//! to the non-VR "player".
+
+use bevy::{
+    prelude::*,
+    render::render_resource::{Extent3d, TextureDimension, TextureFormat},
+};
+use bevy_flycam::{FlyCam, MovementSettings};
+
+/// Creates a colorful test pattern
+fn uv_debug_texture() -> Image {
+    const TEXTURE_SIZE: usize = 8;
+
+    let mut palette: [u8; 32] = [
+        255, 102, 159, 255, 255, 159, 102, 255, 236, 255, 102, 255, 121, 255, 102, 255, 102, 255,
+        198, 255, 102, 198, 255, 255, 121, 102, 255, 255, 236, 102, 255, 255,
+    ];
+
+    let mut texture_data = [0; TEXTURE_SIZE * TEXTURE_SIZE * 4];
+    for y in 0..TEXTURE_SIZE {
+        let offset = TEXTURE_SIZE * y * 4;
+        texture_data[offset..(offset + TEXTURE_SIZE * 4)].copy_from_slice(&palette);
+        palette.rotate_right(4);
+    }
+
+    Image::new_fill(
+        Extent3d {
+            width: TEXTURE_SIZE as u32,
+            height: TEXTURE_SIZE as u32,
+            depth_or_array_layers: 1,
+        },
+        TextureDimension::D2,
+        &texture_data,
+        TextureFormat::Rgba8UnormSrgb,
+    )
+}
+
+#[derive(Component)]
+/// A unique identifier for the non-VR flycam
+pub struct Shape;
+
+pub fn setup(
+    mut commands: Commands,
+    asset_server: Res<AssetServer>,
+    mut meshes: ResMut<Assets<Mesh>>,
+    mut images: ResMut<Assets<Image>>,
+    mut materials: ResMut<Assets<StandardMaterial>>,
+    mut movement_settings: ResMut<MovementSettings>,
+) {
+    let transform =
+        Transform::from_xyz(3., 100., 400.).looking_at(Vec3::new(0.0, 0.3, 0.0), Vec3::Y);
+    commands.spawn((
+        Camera3dBundle {
+            transform,
+            ..default()
+        },
+        FlyCam,
+        EnvironmentMapLight {
+            diffuse_map: asset_server.load("environment_maps/pisa_diffuse_rgb9e5_zstd.ktx2"),
+            specular_map: asset_server.load("environment_maps/pisa_specular_rgb9e5_zstd.ktx2"),
+        },
+    ));
+
+    let material = materials.add(StandardMaterial {
+        base_color_texture: Some(images.add(uv_debug_texture())),
+        ..default()
+    });
+    let mesh = meshes.add(
+        shape::Icosphere {
+            radius: 1.0,
+            subdivisions: 10,
+        }
+        .try_into()
+        .unwrap(),
+    );
+    commands.spawn((
+        PbrBundle {
+            mesh,
+            material,
+            transform,
+            ..default()
+        },
+        Shape,
+    ));
+
+    movement_settings.speed = 100.0;
+}
+
+pub fn move_shape_with_camera(
+    mut shape: Query<&mut Transform, (With<Shape>, Without<FlyCam>)>,
+    camera: Query<&Transform, (Without<Shape>, With<FlyCam>)>,
+) {
+    *shape.single_mut() = *camera.single();
+}
